product, varitaion, category, brand entities, usecases interface, repository interface and Sequelize models

+++ src/application/impl/user/delete-user.ts
@@ -1,4 +1,5 @@
 import { ReadUser } from "../../../domain/entity/user";
+import { NotFoundError } from "../../../domain/exceptions/exceptions";
 import { UserRepository } from "../../../domain/repository/interface/user-repository";
 import { DeleteUserUseCase } from "../../usecases/user/delete-user-usecase";
 
@@ -9,6 +10,8 @@ export class DeleteUserImpl implements DeleteUserUseCase {
   }
 
   async execute(id: number): Promise<void> {
+    const user = await this.userRepository.findOne(id);
+    if (!user) throw new NotFoundError("User");
     await this.userRepository.deleteOne(id);
   }
 }
